Updated files. Major refactor and everything.

+++ Pigeonhole/minimizers.h
@@ -9,9 +9,9 @@
 
 extern vector<pair<string, int> > alphabetRef;
 
-unsigned long long extractRanking(string kmer);
+unsigned long int extractRanking(string kmer);
 uint64_t inthash_64(uint64_t key, uint64_t mask);
-unsigned long long getMinimizerRank(string windowSeed, int q, int windowSize);
+unsigned int getMinimizerRank(string windowSeed, int q, int windowSize);
 //multimap<vector<int>, int> generateMinimizers(string stringDNA, int q, int w, int m);
 
 #endif //MP_PIGEONHOLE_MINIMIZERS_H
